cv-search-service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.24.0
pydantic>=2.0.0
sentence-transformers>=2.2.2
neo4j>=5.8.0
qdrant-client>=1.5.0
httpx>=0.24.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
pyjwt>=2.8.0
slowapi>=0.1.9
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-jaeger>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-httpx>=0.41b0
prometheus-client>=0.17.0